Tidied up log-in routes
