+++ app/brochure/routes/log-in/checkEmail.js
@@ -1,11 +1,11 @@
 var User = require("user");
-var LogInError = require('./logInError');
+var LogInError = require("./logInError");
 
 module.exports = function checkEmail(req, res, next) {
   var email = req.body && req.body.email;
 
   if (!email) return next(new LogInError("NOEMAIL"));
-  
+
   res.locals.email = email;
 
   User.getByEmail(email, function(err, user) {
@@ -16,10 +16,10 @@ module.exports = function checkEmail(req, res, next) {
     if (!user) return next(new LogInError("BADEMAIL"));
 
     // You used to be able to disable your account
-    // but this is no longer possible. Once all 
+    // but this is no longer possible. Once all
     // users with isDisabled:true are removed you
     // can delete this check safely.
-    if (user.isDisabled) return res.redirect('/account/disabled')
+    if (user.isDisabled) return res.redirect("/account/disabled");
 
     res.locals.email = user.email;
     res.locals.validemail = true;
