+++ app/brochure/routes/log-in/index.js
@@ -1,18 +1,13 @@
-var helper = require("helper");
-var Email = helper.email;
-var format = require("url").format;
-var User = require("user");
-var NOTOKEN = "Could not generate a token";
-var config = require("config");
-var generateAccessToken = User.generateAccessToken;
+var BodyParser = require("body-parser");
+var Express = require("express");
+
 var checkToken = require("./checkToken");
+var checkReset = require("./checkReset");
 var checkEmail = require("./checkEmail");
 var checkPassword = require("./checkPassword");
-var LogInError = require("./logInError");
-// Error codes & their corresponding message
-var parse = require("body-parser").urlencoded({ extended: false });
+var errorHandler = require("./errorHandler");
+var parse = BodyParser.urlencoded({ extended: false });
 
-var Express = require("express");
 var form = new Express.Router();
 
 // form.use(require("./rateLimit"));
@@ -38,51 +33,11 @@ form
     res.render("log-in/reset");
   })
 
-  .post(checkEmail, checkReset, errorHandler)
+  .post(parse, checkEmail, checkReset, errorHandler)
 
   .post(function(err, req, res, next) {
     res.render("log-in/reset");
   });
-
-function checkReset(req, res, next) {
-  var user = req.user;
-  var hasPassword = user.passwordHash !== "";
-  var reset = req.body && req.body.reset !== undefined;
-
-  // Some users have not yet set up a password
-  // so we might need to send them a link even
-  // if they did not click the reset button
-  if (!reset && hasPassword) return next();
-
-  sendPasswordResetEmail(user.uid, function(err) {
-    if (err) return next(err);
-
-    res.locals.sent = true;
-    res.locals.hasPassword = hasPassword;
-    res.render("log-in/reset");
-  });
-}
-
-function sendPasswordResetEmail(uid, callback) {
-  var url;
-
-  generateAccessToken(uid, function(err, token) {
-    if (err || !token) return callback(err || new Error(NOTOKEN));
-
-    // The full one-time log-in link to be sent to the user
-    url = format({
-      protocol: "https",
-      host: config.host,
-      pathname: "/log-in",
-      query: {
-        token: token,
-        then: "/account/password/set"
-      }
-    });
-
-    Email.SET_PASSWORD(uid, { url: url }, callback);
-  });
-}
 
 form
   .route("/")
@@ -96,20 +51,7 @@ form
     if (req.body && req.body.reset !== undefined)
       return res.redirect("/log-in/reset");
 
-    console.log("LOCALS:", res.locals);
     res.render("log-in");
   });
 
-function errorHandler(err, req, res, next) {
-  if (!(err instanceof LogInError)) {
-    console.log(err);
-    return next(err);
-  }
-
-  res.locals.error = res.locals[err.code] = true;
-  res.locals.email = req.body && req.body.email;
-  res.status(403);
-
-  next(err);
-}
 module.exports = form;
