+++ app/brochure/views/log-in/reset.html
@@ -39,7 +39,7 @@ input[type="submit"].link:hover {
     <input name="email" value="{{email}}" type="text" autofocus />
   </p>
 
-  <input type="submit" value="Reset your password" />
+  <input type="submit" name="reset"  value="Reset your password" />
 </form>
 {{/sent}}
